# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.AvaialbilityResponse:
    properties:
      available:
        type: boolean
      availableUnits:
        type: integer
      endAt:
        type: string
      equipmentID:
        type: integer
      startAt:
        type: string
    type: object
  dto.EquipmentCreateRequest:
    properties:
      category:
        type: string
      dailyRate:
        format: float64
        type: number
      description:
        type: string
      name:
        type: string
      quantity:
        type: integer
      salePrice:
        format: float64
        type: number
      serials:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  dto.EquipmentDetailResponse:
    properties:
      availableUnits:
        type: integer
      category:
        type: string
      createdAt:
        type: string
      dailyRate:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      salePrice:
        type: string
      serials:
        items:
          type: string
        type: array
      type:
        type: string
      updatedAt:
        type: string
    type: object
  dto.EquipmentResponse:
    properties:
      category:
        type: string
      createdAt:
        type: string
      dailyRate:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      salePrice:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  dto.EquipmentUpdateRequest:
    properties:
      category:
        type: string
      dailyRate:
        format: float64
        type: number
      description:
        type: string
      name:
        type: string
      quantity:
        type: integer
      salePrice:
        format: float64
        type: number
      type:
        type: string
    type: object
  dto.InvoiceResponse:
    properties:
      amount:
        format: float64
        type: number
      invoiceNumber:
        type: string
      invoiceStatus:
        type: string
      issuedAt:
        type: string
      orderID:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.MyResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  dto.OrderCreateRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.OrderItemCreateRequest'
        type: array
    type: object
  dto.OrderItemCreateRequest:
    properties:
      endAt:
        type: string
      equipmentID:
        type: integer
      itemType:
        type: string
      quantity:
        type: integer
      reservationID:
        type: integer
      startAt:
        type: string
    type: object
  dto.OrderItemResponse:
    properties:
      endAt:
        type: string
      equipmentID:
        type: integer
      equipmentUnitID:
        type: integer
      id:
        type: integer
      itemType:
        type: string
      lineTotal:
        format: float64
        type: number
      quantity:
        type: integer
      startAt:
        type: string
      unitPrice:
        format: float64
        type: number
    type: object
  dto.OrderResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.OrderItemResponse'
        type: array
      orderType:
        type: string
      status:
        type: string
      totalAmount:
        format: float64
        type: number
      userId:
        type: integer
    type: object
  dto.OrderStatusUpdateRequest:
    properties:
      status:
        type: string
    type: object
  dto.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.RentalBookRequest:
    properties:
      endAt:
        type: string
      equipmentID:
        type: integer
      startAt:
        type: string
    type: object
  dto.RentalBookResponse:
    properties:
      endAt:
        type: string
      equipmentId:
        type: integer
      equipmentUnit:
        type: integer
      estimatedCost:
        format: float64
        type: number
      reservationID:
        type: integer
      startAt:
        type: string
      status:
        type: string
    type: object
  dto.RentalCalculateRequest:
    properties:
      endAt:
        type: string
      equipmentId:
        type: integer
      mode:
        type: string
      startAt:
        type: string
    type: object
  dto.RentalCalculateResponse:
    properties:
      amount:
        format: float64
        type: number
      endAt:
        type: string
      equipmentId:
        type: integer
      mode:
        type: string
      startAt:
        type: string
    type: object
  dto.ReqisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: Equipment rental and sales API.
  title: Rental API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReqisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a user
      tags:
      - auth
  /equipment:
    get:
      parameters:
      - description: Equipment type
        in: query
        name: type
        type: string
      - description: Equipment category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EquipmentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List equipment
      tags:
      - equipment
    post:
      consumes:
      - application/json
      parameters:
      - description: Create equipment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EquipmentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EquipmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create equipment
      tags:
      - equipment
  /equipment/{id}:
    get:
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EquipmentDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get equipment by id
      tags:
      - equipment
    put:
      consumes:
      - application/json
      parameters:
      - description: Equipment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update equipment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EquipmentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EquipmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update equipment
      tags:
      - equipment
  /orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.OrderResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Create order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OrderCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /orders/{id}/invoice:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get order invoice
      tags:
      - orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OrderStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - orders
  /rentals/availability:
    get:
      parameters:
      - description: Equipment ID
        in: query
        name: equipmentId
        required: true
        type: integer
      - description: Start datetime (RFC3339)
        in: query
        name: startAt
        required: true
        type: string
      - description: End datetime (RFC3339)
        in: query
        name: endAt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvaialbilityResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check rental availability
      tags:
      - rentals
  /rentals/book:
    post:
      consumes:
      - application/json
      parameters:
      - description: Book rental request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RentalBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RentalBookResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Book rental
      tags:
      - rentals
  /rentals/calculate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Rental calculation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RentalCalculateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RentalCalculateResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Calculate rental amount
      tags:
      - rentals
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MyResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MyResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"